Replace zero E-values
I used an alignment between 535 lysines and 534 lysines containing a single
internal arginine to empirically determine that 1e-180 is the smallest E-value
that BLASTP 2.2.28+ will report before rounding down to zero. I therefore
heuristically replace all zeros with 1e-181. This is a modification of the
OrthoMCL heuristic, which replaces zeros with a value 10 less than the minimum
non-zero E-value observed within the BLASTP results for a given run. A better
approach would be to recompute all E-values that have been rounded to zero, but
I'm currently not quite sure how BLAST+ determines the effective length for
each query.

Signed-off-by: Ted Gibbons <ted.codes@gmail.com>