Added a new file for formatting the COG fasta files
The COG databases available through the NCBI FTP site contains FASTA files with
complete transcriptomes, and whog/twog files relating subsets of these
sequences. I have added a program to the project (formatCOGfasta.py) that uses
the whog/twog files to extract the annotated sequences from the complete
transcriptome files and reformats the sequence IDs in a way that is compatible
with both OrthoMCL and CEGMA, and contains most of the information contained
in the whog/twog files. I also continued to tighten up the code for
blast2graph.py.

Signed-off-by: Ted Gibbons <ted.codes@gmail.com>