Added file to consolidate many clusterings into an R dataframe
This program takes an output prefix, and an arbitrarily large number of MCL
output files containing clusters of CEGMA (or custom-formatted COG) sequences.
The format of the file names is very restrictive, as this program was meant to
be run as the penultimate step in a larger automated pipeline. It outputs two
text tables that are meant to be easily read into R dataframes for creating
summary plots.

Signed-off-by: trgibbons <ted.codes@gmail.com>