Glue script for fully-automated CEGMA clustering pipeline
This BASH script brings the number of languages in my pipeline to three and
the underscore-delimited name brings the number of script-naming conventions
to three as well. This script takes only a single argument, a six-digit integer
encoding a scheme for fragmenting the sequences in the CEGMA database. Using
this number, the pipeline generates four test data sets, two all-v-all BLASTp
files for each data set, two sets of four graphs for each BLASTp file, 40
clusterings for graph, and then two PDFs with heatmaps for each set of four
graphs.

Signed-off-by: trgibbons <ted.codes@gmail.com>