Updated title of legend
Realized that the legend had been incorrectly listing the color range as
indicating a linear relationship, when actually it was showing a logarithmic
relationship.  Made this change back in March, but I've since switched to using
stacked barcharts, which I find to be more useful, so I'm just getting around
to committing this change.

Signed-off-by: Ted Gibbons <ted.codes@gmail.com>