Added a single script to generate the ECK database
As this project evolved, I switched from using the complete KOG database, to
the CEGMA database, to a custom database created by restoring the inparalogs to
CEGMA ("CEGMA KOGs"), and finally to an extension of the CEGMA KOGs that
includes the sequences annotated for the second version of CEGMA, which I call
the "Expanded CEGMA KOGs" database.

I created each iteration by using custom scripts to modify the previous
version, making it difficult for someone else to reproduce. This script should
simplify the task by fetching the files I used from the internet and combining
them into the ECK database.

I also took the opportunity to add some extra information to the FASTA headers.

Signed-off-by: Ted Gibbons <tgibbons@Sleipnir.local>