Pipeline no longer uses novel method for supplementing BLAST-rounded E-values
OrthoMCL replaces all "zero" E-values with a number whose exponent was one less
than the smallest E-value observed in the remainder of the data set. It does
this before inter-organism normalization, which means extremely small E-values
can be rounded to zero, heuristically set to a value just below the edge of the
truncated distribution, and then shifted up into that distribution during
normalization. I came up with a way to prevent this, but it appeared to
actually hinder the performance of the -log10(E-value) metric in practice, so I
switched to using the OrthoMCL method.

Signed-off-by: Ted Gibbons <tgibbons@Sleipnir.local>